Address all 17 review items on e2e test
- Split shared.ts into config, runner, accounts, http402, sdk modules
- Use role-scoped presets (createMerchantClient, createArbiterClient)
  with createX402r for payer (needs full freeze access)
- Inline evidence bytecode (no cross-repo dependency)
- Tighten assertions: exact PAYMENT_AMOUNT checks, approvedAmount,
  evidence submitter addresses, finalRefundStatus
- runner.fail() now throws to prevent cascading failures
- Use runner.assert() throughout instead of verbose if/else
- Use chainConfig.usdc instead of hardcoded USDC_ADDRESS
- Properly type FacilitatorClientLike (no unknown/any)
- Remove unused exports, dead code, HTTP402PaymentResult fields
- Fix distributeFees: check accumulated fees before calling
- Rewrite README to match actual SDK usage
- 21/21 steps pass on Base Sepolia
