+++ examples/e2e-test/sdk.ts
@@ -1,0 +1,169 @@
+/**
+ * SDK instance creation, operator deployment, and evidence contract bytecode.
+ */
+
+import {
+  deployMarketplaceOperator,
+  refundRequestEvidenceAbi,
+} from '@x402r/core'
+import {
+  type ArbiterClient,
+  createArbiterClient,
+  createMerchantClient,
+  createX402r,
+  type MerchantClient,
+  type X402r,
+} from '@x402r/sdk'
+import { type Address, createWalletClient, http } from 'viem'
+import { baseSepolia } from 'viem/chains'
+
+import type { E2EAccounts } from './accounts.js'
+import { RPC_URL, waitForTx } from './config.js'
+import type { StepRunner } from './runner.js'
+
+// ============ Evidence Contract Bytecode (inline) ============
+// Source: x402r-contracts/out/RefundRequestEvidence.sol/RefundRequestEvidence.json → bytecode.object
+// Inlined to avoid cross-repo filesystem dependency at runtime.
+
+const EVIDENCE_BYTECODE: `0x${string}` =
+  '0x60a034607657601f61120138819003918201601f19168301916001600160401b03831184841017607a5780849260209460405283398101031260765751600160' +
+  '0160a01b03811690819003607657608052604051611172908161008f82396080518181816103fb0152818161084c01526110950152f35b5f80fd5b634e487b7160' +
+  'e01b5f52604160045260245ffdfe6080806040526004361015610012575f80fd5b5f3560e01c9081633216d6ab1461087b5750806353c3220814610837578063872f' +
+  '69051461034a57806388d1117e146102b45763c622ee7d14610053575f80fd5b346102705736600319016101c081126102705761018013610270576101a43561007a' +
+  '610bec565b50600460206001600160a01b0361008f610a03565b1660405192838092637340e25560e11b82525afa90811561027c57610184916020915f91610287575b' +
+  '5060405163063a70ff60e01b815292839182906100d760048301610a76565b6001600160a01b03165afa90811561027c575f91610246575b506040516020810191825261' +
+  '018435604082015260408152610113606082610a35565b519020805f52600160205260405f2054821015610237575f525f60205260405f20905f5260205260405f2060016040519161014d83610a19565b65ffffffffffff8154838060a01b038116855261017360ff8260a01c1660208701610b90565b60a81c16604084015201906040515f9280549061018f82610b9c565b808452916001811690811561020f57506001146101d5575b50506101b8816101d1940382610a35565b606082015260405191829160208352602083019061099e565b0390f35b9093505f5260205f205f905b8482106101f95750810160200192506101b8816101a7565b60018160209254838587010152019101906101e1565b6101d1965084925060209150926101b89360ff191682840152151560051b82010194506101a7565b634e23d03560e01b5f5260045ffd5b90506020813d602011610274575b8161026160209383610a35565b8101031261027057515f6100f0565b5f80fd5b3d9150610254565b6040513d5f823e3d90fd5b6102a79150823d84116102ad575b61029f8183610a35565b810190610a57565b5f6100b8565b503d610295565b346102705736600319016101e081126102705761018013610270576102e36101c4356101a43561018435610c38565b9060405190604082016040835281518091526060830190602060608260051b8601019301915f905b82821061031f578580868960208301520390f35b9091929360208061033c600193605f198a8203018652885161099e565b96019201920190929161030b565b346102705736600319016101c08112610270576101801361027057610184356101a43567ffffffffffffffff811161027057366023820112156102705780600401359167ffffffffffffffff83116102705760248201916024843692010111610270576001600160a01b036103bd610a03565b1615610828578215610819576103d1611048565b60405163500431a960e11b81526103ea60048201610a76565b61018481018390526020816101a4817f00000000000000000000000000000000000000000000000000000000000000006001600160a01b03165afa90811561027c575f916107de575b50156107cf57600460206001600160a01b0361044d610a03565b1660405192838092637340e25560e11b82525afa90811561027c57610184916020915f916107b2575b5060405163063a70ff60e01b8152928391829061049560048301610a76565b6001600160a01b03165afa90811561027c575f91610780575b5060405160208101918252836040820152604081526104ce606082610a35565b519020805f52600160205260405f205491604051956104ec87610a19565b338752602087016104fd8382610b90565b6040880165ffffffffffff42168152601f19601f840116986040519261052660208c0185610a35565b848452848a60208601375f6020868601015260608201938452865f525f60205260405f20885f5260205260405f209160018060a01b039051168254915190600382101561076c5760019465ffffffffffff60a81b905160a81b169264ffffffffff60d81b16179060ff60a01b9060a01b161717815501905180519067ffffffffffffffff82116107585781906105bc8454610b9c565b601f81116106ff575b50602090601f831160011461069c575f92610691575b50508160011b915f199060031b1c19161790555b6001840180851161067d577f84f5dbe467ce8287f18d3550c85b5a1c85ba9a3dbebcfcd280cf0af05661489c966106516102209484935f97885260016020526040882055604051986106408a610a76565b6101808a01526101a0890190610991565b6102006101c08801528161020088015283870137840101526101e08201526102208133948101030190a2005b634e487b7160e01b5f52601160045260245ffd5b015190508a806105db565b5f8581528281209350601f198516905b8181106106e757509084600195949392106106cf575b505050811b0190556105ef565b01515f1960f88460031b161c191690558a80806106c2565b929360206001819287860151815501950193016106ac565b828111156105c557909150835f5260205f20601f840160051c9060208510610750575b849392601f0160051c82900391015f5b8281106107405750506105c5565b5f81830155859450600101610732565b5f9150610722565b634e487b7160e01b5f52604160045260245ffd5b634e487b7160e01b5f52602160045260245ffd5b90506020813d6020116107aa575b8161079b60209383610a35565b810103126102705751856104ae565b3d915061078e565b6107c99150823d84116102ad5761029f8183610a35565b87610476565b6377840b9360e01b5f5260045ffd5b90506020813d602011610811575b816107f960209383610a35565b81010312610270575180151581036102705785610433565b3d91506107ec565b636d91186360e01b5f5260045ffd5b63ccea9e6f60e01b5f5260045ffd5b34610270575f366003190112610270576040517f00000000000000000000000000000000000000000000000000000000000000006001600160a01b03168152602090f35b346102705736600319016101a081126102705761018013610270576020816004816001600160a01b036108ac610a03565b637340e25560e11b8352165afa90811561027c57610184916020915f91610974575b5060405163063a70ff60e01b815292839182906108ed60048301610a76565b6001600160a01b03165afa90811561027c575f91610942575b506040516020810191825261018435604082015260408152610929606082610a35565b5190205f526001602052602060405f2054604051908152f35b90506020813d60201161096c575b8161095d60209383610a35565b81010312610270575181610906565b3d9150610950565b61098b9150823d84116102ad5761029f8183610a35565b836108ce565b90600382101561076c5752565b80516001600160a01b0316825260208082015160a093926060916109c59084860190610991565b65ffffffffffff60408201511660408501520151608060608401528051918291826080860152018484015e5f828201840152601f01601f1916010190565b6004356001600160a01b03811681036102705790565b6080810190811067ffffffffffffffff82111761075857604052565b90601f8019910116810190811067ffffffffffffffff82111761075857604052565b9081602091031261027057516001600160a01b03811681036102705790565b6004356001600160a01b038116908190036102705781526024356001600160a01b038116908190036102705760208201526044356001600160a01b038116908190036102705760408201526064356001600160a01b038116908190036102705760608201526084356001600160781b03811680910361027057608082015260a43565ffffffffffff81168091036102705760a082015260c43565ffffffffffff81168091036102705760c082015260e43565ffffffffffff81168091036102705760e08201526101043561ffff8116809103610270576101008201526101243561ffff811680910361027057610120820152610144356001600160a01b038116908190036102705761014082015261016061016435910152565b600382101561076c5752565b90600182811c92168015610bca575b6020831014610bb657565b634e487b7160e01b5f52602260045260245ffd5b91607f1691610bab565b67ffffffffffffffff81116107585760051b60200190565b60405190610bf982610a19565b606080835f81525f60208201525f60408201520152565b8051821015610c245760209160051b010190565b634e487b7160e01b5f52603260045260245ffd5b919291600435906001600160a01b03821680830361027057604051637340e25560e11b815290602082600481845afa91821561027c575f92611027575b5060405163063a70ff60e01b8152938190036102705760048401526024356001600160a01b0381168103610270576001600160a01b0390811660248501526044359081168103610270576001600160a01b0390811660448501526064359081168103610270576001600160a01b031660648401526084356001600160781b0381169081900361027057608484015260a43565ffffffffffff811681036102705765ffffffffffff1660a484015260c43565ffffffffffff811681036102705765ffffffffffff1660c484015260e43565ffffffffffff811681036102705765ffffffffffff1660e48401526101043561ffff811681036102705761ffff166101048401526101243561ffff811681036102705761ffff16610124840152610144356001600160a01b0381168103610270576001600160a01b039081166101448501526101648035908501526020918491610184918391165afa91821561027c575f92610ff3575b506040519060208201928352604082015260408152610dfc606082610a35565b51902090815f52600160205260405f205493848210801590610feb575b610fb0578185039085821161067d5780821015610fa85750905b610e3c82610bd4565b92610e4a6040519485610a35565b828452601f19610e5984610bd4565b015f5b818110610f915750505f5b838110610e7657505050509190565b815f525f60205260405f2081840180851161067d575f5260205260405f20600160405191610ea383610a19565b65ffffffffffff8154838060a01b0381168552610ec960ff8260a01c1660208701610b90565b60a81c16604084015201604051905f90805490610ee582610b9c565b8085529160018116908115610f6a5750600114610f31575b505090610f108160019594930382610a35565b6060820152610f1f8288610c10565b52610f2a8187610c10565b5001610e67565b5f908152602081209092505b818310610f545750508101602001610f1082610efd565b6001816020925483868801015201920191610f3d565b60ff191660208087019190915292151560051b85019092019250610f109150839050610efd565b602090610f9c610bec565b82828901015201610e5c565b905090610e33565b505050604051610fc1602082610a35565b5f81525f805b818110610fd45750509190565b602090610fdf610bec565b82828601015201610fc7565b508015610e19565b9091506020813d60201161101f575b8161100f60209383610a35565b810103126102705751905f610ddc565b3d9150611002565b61104191925060203d6020116102ad5761029f8183610a35565b905f610c75565b6024356001600160a01b0381169081810361027057503314611138576044356001600160a01b0381169081810361027057503314611133576040516324132b9760e21b81526020816004817f00000000000000000000000000000000000000000000000000000000000000006001600160a01b03165afa90811561027c575f916110f1575b506001600160a01b031633146110ec5763150e886360e31b5f5260045ffd5b600290565b90506020813d60201161112b575b8161110c60209383610a35565b8101031261027057516001600160a01b0381168103610270575f6110cd565b3d91506110ff565b600190565b5f9056fea26469706673582212205965c66ee68aaf0dca8eadeedfe14479caf4d5e0d6f7dd7dba5eae9f830c506864736f6c63430008210033'
+
+// ============ Types ============
+
+interface DeployResult {
+  operatorAddress: Address
+  escrowPeriodAddress: Address
+  freezeAddress: Address | null
+  refundRequestAddress: Address
+  refundInEscrowConditionAddress: Address
+  refundRequestEvidenceAddress: Address
+  feeCalculatorAddress: Address | null
+  summary: {
+    newCount: number
+    existingCount: number
+    txHashes: `0x${string}`[]
+  }
+}
+
+interface SDKInstances {
+  payer: X402r
+  merchant: MerchantClient
+  arbiter: ArbiterClient | undefined
+}
+
+// ============ Operator Deployment ============
+
+export async function deployTestOperator(
+  accounts: E2EAccounts,
+  runner: StepRunner,
+): Promise<DeployResult> {
+  const { payerAccount, arbiterAccount, publicClient, chainId } = accounts
+
+  // Fresh wallet client to avoid stale nonce from prior funding transactions
+  const deployWallet = createWalletClient({
+    account: payerAccount,
+    chain: baseSepolia,
+    transport: http(RPC_URL),
+  })
+
+  const deployResult = await deployMarketplaceOperator(
+    deployWallet,
+    publicClient,
+    {
+      chainId,
+      feeRecipient: payerAccount.address,
+      arbiter: (arbiterAccount ?? payerAccount).address,
+      escrowPeriodSeconds: 604800n,
+      freezeDurationSeconds: 259200n,
+      operatorFeeBps: 100n,
+    },
+  )
+
+  runner.log(`Operator: ${deployResult.operatorAddress}`)
+  runner.pass(
+    'Deploy operator',
+    deployResult.summary.txHashes.length > 0
+      ? deployResult.summary.txHashes[0]
+      : undefined,
+  )
+
+  // Deploy RefundRequestEvidence bound to the per-arbiter RefundRequest
+  // Wait for RPC nonce to catch up after multicall batch deploy
+  await new Promise((r) => setTimeout(r, 3000))
+  runner.log('Deploying RefundRequestEvidence...')
+  const evidenceWallet = createWalletClient({
+    account: payerAccount,
+    chain: baseSepolia,
+    transport: http(RPC_URL),
+  })
+  const evidenceDeployTx = await evidenceWallet.deployContract({
+    abi: refundRequestEvidenceAbi,
+    bytecode: EVIDENCE_BYTECODE,
+    args: [deployResult.refundRequestAddress],
+    account: payerAccount,
+    chain: baseSepolia,
+  })
+  await waitForTx(publicClient, evidenceDeployTx)
+  const evidenceReceipt = await publicClient.getTransactionReceipt({
+    hash: evidenceDeployTx,
+  })
+  const evidenceAddress = evidenceReceipt.contractAddress as Address
+  runner.log(`  RefundRequestEvidence: ${evidenceAddress}`)
+  runner.pass('Deploy RefundRequestEvidence', evidenceDeployTx)
+
+  return {
+    operatorAddress: deployResult.operatorAddress,
+    escrowPeriodAddress: deployResult.escrowPeriodAddress,
+    freezeAddress: deployResult.freezeAddress,
+    refundRequestAddress: deployResult.refundRequestAddress,
+    refundInEscrowConditionAddress: deployResult.refundInEscrowConditionAddress,
+    refundRequestEvidenceAddress: evidenceAddress,
+    feeCalculatorAddress: deployResult.feeCalculatorAddress,
+    summary: deployResult.summary,
+  }
+}
+
+// ============ SDK Instance Creation (role-scoped presets) ============
+
+export function createSDKInstances(
+  accounts: E2EAccounts,
+  deployResult: DeployResult,
+): SDKInstances {
+  const { publicClient, payerWallet, merchantWallet, arbiterWallet, chainId } =
+    accounts
+  const opAddr = deployResult.operatorAddress
+  const refundRequestAddress = deployResult.refundRequestAddress
+  const freezeAddr = deployResult.freezeAddress ?? undefined
+  const escrowPeriodAddr = deployResult.escrowPeriodAddress
+
+  const commonConfig = {
+    publicClient,
+    operatorAddress: opAddr,
+    chainId,
+    refundRequestAddress,
+    refundRequestEvidenceAddress: deployResult.refundRequestEvidenceAddress,
+    escrowPeriodAddress: escrowPeriodAddr,
+    freezeAddress: freezeAddr,
+  }
+
+  // Use createX402r for payer (needs full freeze access — payer initiates freezes)
+  const payer = createX402r({
+    ...commonConfig,
+    walletClient: payerWallet,
+  })
+  const merchant = createMerchantClient({
+    ...commonConfig,
+    walletClient: merchantWallet,
+  })
+  const arbiter = arbiterWallet
+    ? createArbiterClient({ ...commonConfig, walletClient: arbiterWallet })
+    : undefined
+
+  return { payer, merchant, arbiter }
+}
